test(functional): cover additional moe perf recipes
