+++ tests/functional_tests/launch_scripts/gb200/active/L0_Launch_recipes_moe_perf_proxies.sh
@@ -1,13 +1,13 @@
 #!/bin/bash
 # GPU_COUNT=x8
-# CI_TIMEOUT=30
+# CI_TIMEOUT=60
 # Copyright (c) 2026, NVIDIA CORPORATION.  All rights reserved.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
 # You may obtain a copy of the License at
 #
-#     http://www.apache.org/licenses/LICENSE-2.0
+# http://www.apache.org/licenses/LICENSE-2.0
 #
 # Unless required by applicable law or agreed to in writing, software
 # distributed under the License is distributed on an "AS IS" BASIS,
@@ -27,15 +27,30 @@ export NCCL_NVLS_ENABLE=0
 export NUM_OF_HYBRID_EP_RANKS_PER_NVLINK_DOMAIN=8
 export NUM_OF_TOKENS_PER_CHUNK_COMBINE_API=128
 export NVLINK_DOMAIN_SIZE=72
+export NVTE_BWD_LAYERNORM_SM_MARGIN=20
 export NVTE_CUTEDSL_FUSED_GROUPED_MLP=1
+export NVTE_FWD_LAYERNORM_SM_MARGIN=20
 export PYTORCH_CUDA_ALLOC_CONF="expandable_segments:True"
+export TOKENIZERS_PARALLELISM=False
 export TORCH_NCCL_AVOID_RECORD_STREAMS=1
 export TORCH_NCCL_HIGH_PRIORITY=1
 export USE_MNNVL=1
 
-uv run python -m torch.distributed.run --nproc_per_node=8 --nnodes=1 -m coverage run \
-  --data-file=/opt/Megatron-Bridge/.coverage --source=/opt/Megatron-Bridge/ --parallel-mode \
-  -m pytest -o log_cli=true -o log_cli_level=INFO -v -s -x -m "not pleasefixme" --tb=short -rA \
-  tests/functional_tests/test_groups/recipes/test_qwen3_moe_perf_proxy.py::TestQwen3MoePerfProxy::test_gb200_fp8mx
+run_proxy() {
+  uv run python -m torch.distributed.run --nproc_per_node=8 --nnodes=1 -m coverage run \
+    --data-file=/opt/Megatron-Bridge/.coverage --source=/opt/Megatron-Bridge/ --parallel-mode \
+    -m pytest -o log_cli=true -o log_cli_level=INFO -v -s -x -m "not pleasefixme" --tb=short -rA "$1"
+}
+
+run_proxy \
+  tests/functional_tests/test_groups/recipes/test_moe_perf_proxies.py::TestQwen3MoePerfProxy::test_gb200_fp8mx
+run_proxy \
+  tests/functional_tests/test_groups/recipes/test_moe_perf_proxies.py::TestAdditionalMoePerfProxies::test_gpt_oss_120b_gb200_fp8mx
+
+export NVTE_ALLOW_NONDETERMINISTIC_ALGO=0
+export NVTE_NORM_BWD_USE_CUDNN=1
+export NVTE_NORM_FWD_USE_CUDNN=1
+run_proxy \
+  tests/functional_tests/test_groups/recipes/test_moe_perf_proxies.py::TestAdditionalMoePerfProxies::test_deepseek_v3_gb200_fp8mx
 
 coverage combine -q
